move logic out of binary
